Add legend to GlobeMap and move MapLegend component to a new file
